+++ src/components/Footers/FooterBlackSocial.js
@@ -210,14 +210,14 @@ function FooterBlackSocial() {
               href="https://www.invisionapp.com?ref=creativetim"
               target="_blank"
             >
-              Invision
+             Jarb29
             </a>
             . Coded by{" "}
             <a
               href="https://www.creative-tim.com?ref=nuk-pro-react-footer-black-social"
               target="_blank"
             >
-              Creative Tim
+             Jarb29
             </a>
             .
           </div>